Run threads longer
The idea is to kick off a new thread if we are running fewer threads
than the requested limit, but otherwise continue recursion within the
same thread.  This would result in starting fewer threads, and running
them longer.

This so far is strangely running faster than the threaded implementation
previously did, but no faster than the single-threaded implementation.