Abandon inferior encodings early
Since the largest encoding (int) is the "canonical" one, as soon as we
determine that we are creating an smaller encding we can abandon the
encording determined by the start cube and rotation.

We can also build our encoding with shifts+ands as we go through the
depth-first traversal and don't need to create the encoding afterward
with the encoding's ordered set of cubes.