Increase CPU utilization with probabilistic halt check
The main loop previously polled the halt pipe twice per invocation.
With random 0.1% (1/1000) chance of a halt check instead, we can still
halt the program quickly while increasing the per-process CPU utlization
from about 90% to 98%.  Previously, a lot of time was spent waiting for
the Manager.Pipe.poll() call to return.

The halt process was also simplified a bit with a refactor to use
exceptions instead.  This makes enqueueing Polycube instances for
writing to disk more clean and easy to read.