Actual fix for resuming, and undo previous fix attempt
The two problems that were fixed:
- worker threads were not getting the halted flag set right away, so
  they'd continue work on polycubes when they should just be returning
  them to be written to disk
- worker threads now use a single Polycube instance for their recursive
  evaluation -- this means a copy of the original unmodified Polycube
  needs to be kept in case the worker is halted

Another problem is strange output by the flate2::write::GzEncoder.  When
writing several tens of thousands of lines of comma-delimited integers,
some invalid data would be written that would be decompressed as "--401"
and "-2-209" or similar.  The only fix that works for now is to use a
compression level of "none".