Add --write-found... and --begin-from... options
These options allow the final set of found polycubes to be written to
disk, and then using a set of found polycubes as a starting point to
find larger polycubes.

This changes the way polycubes are written to disk: only the canonical
encoding integer is written for each polycube.  To save a little space,
the encoding integers are written in base94.  To make the file a little
easier to work with (to allow sorting before compression) the files are
not being written in binary mode.